Completely modularise argument parsing
This is also a pre-requisite for adding support for MacOS
