+++ src/austin.c
@@ -27,6 +27,7 @@
 #include <sys/types.h>
 #include <unistd.h>
 
+#include "argparse.h"
 #include "austin.h"
 #include "error.h"
 #include "logging.h"
@@ -39,21 +40,7 @@
 #include "py_thread.h"
 
 
-#define DEFAULT_SAMPLING_INTERVAL    100
-
-const char SAMPLE_FORMAT_NORMAL[]      = ";%s (%s);L%d";
-const char SAMPLE_FORMAT_ALTERNATIVE[] = ";%s (%s:%d)";
-
 static ctime_t t_sample;  // Checkmark for sampling duration calculation
-
-// Globals for command line arguments
-static ctime_t t_sampling_interval = DEFAULT_SAMPLING_INTERVAL;
-static pid_t   attach_pid          = 0;
-static int     exclude_empty       = 0;
-static int     sleepless           = 0;
-static char *  format              = (char *) SAMPLE_FORMAT_NORMAL;
-
-static int exec_arg = 0;
 
 
 // ---- HELPERS ---------------------------------------------------------------
@@ -124,17 +111,6 @@ _py_proc__sample(py_proc_t * py_proc) {
 }
 
 
-// ----------------------------------------------------------------------------
-static int
-strtonum(char * str, long * num) {
-  char * p_err;
-
-  *num = strtol(str, &p_err, 10);
-
-  return  (p_err == str || *p_err != 0) ? 1 : 0;
-}
-
-
 // ---- SIGNAL HANDLING -------------------------------------------------------
 
 static int interrupt = 0;
@@ -147,230 +123,13 @@ signal_callback_handler(int signum)
 }
 
 
-// ---- ARGUMENT PARSING ------------------------------------------------------
-
-#if defined(__linux__)
-#include <argp.h>
-
-const char * argp_program_version = PROGRAM_NAME " " VERSION;
-
-const char * argp_program_bug_address = \
-  "<https://github.com/P403n1x87/austin/issues>";
-
-static const char * doc = "Austin -- A frame stack sampler for Python.";
-
-static struct argp_option options[] = {
-  {
-    "interval",     'i', "n_usec",      0,
-    "Sampling interval (default is 500 usec)."
-  },
-  {
-    "alt-format",   'a', NULL,          0,
-    "alternative collapsed stack sample format."
-  },
-  {
-    "exclude-empty",'e', NULL,          0,
-    "do not output samples of threads with no frame stacks."
-  },
-  {
-    "sleepless",    's', NULL,          0,
-    "suppress idle samples."
-  },
-  {
-    "pid",          'p', "PID",         0,
-    "The the ID of the process to which Austin should attach."
-  },
-  {0}
-};
-
-
-// ----------------------------------------------------------------------------
-static int
-parse_opt (int key, char *arg, struct argp_state *state)
-{
-  if (state->argc == 1) {
-    argp_state_help(state, stdout, ARGP_HELP_USAGE);
-    exit(0);
-  }
-
-  // Consume all the remaining arguments if the next one is not an option so
-  // that they can be passed to the command to execute
-  if ((state->next == 0 && state->argv[1][0] != '-')
-  ||  (state->next > 0 && state->next < state->argc && state->argv[state->next][0] != '-')
-  ) {
-    exec_arg = state->next == 0 ? 1 : state->next;
-    state->next = state->argc;
-  }
-
-  long l_pid;
-  switch(key) {
-  case 'i':
-    if (strtonum(arg, (long *) &t_sampling_interval) == 1 || t_sampling_interval < 0)
-      argp_error(state, "the sampling interval must be a positive integer");
-    break;
-
-  case 'a':
-    format = (char *) SAMPLE_FORMAT_ALTERNATIVE;
-    break;
-
-  case 'e':
-    exclude_empty = 1;
-    break;
-
-  case 's':
-    sleepless = 1;
-    break;
-
-  case 'p':
-    if (strtonum(arg, &l_pid) == 1 || l_pid <= 0)
-      argp_error(state, "invalid PID.");
-    attach_pid = (pid_t) l_pid;
-    break;
-
-  case ARGP_KEY_ARG:
-  case ARGP_KEY_END:
-    if (attach_pid != 0 && exec_arg != 0)
-      argp_error(state, "the -p option is incompatible with the command argument.");
-    break;
-
-  default:
-    return ARGP_ERR_UNKNOWN;
-  }
-
-  return 0;
-}
-
-
-#elif defined(_WIN32) || defined(_WIN64)
-#include "win/argparse.h"
-
-static const char * help_msg = \
-"Usage: austin [OPTION...] command [ARG...]\n"
-"Austin -- A frame stack sampler for Python.\n"
-"\n"
-"  -a, --alt-format           alternative collapsed stack sample format.\n"
-"  -e, --exclude-empty        do not output samples of threads with no frame\n"
-"                             stacks.\n"
-"  -i, --interval=n_usec      Sampling interval (default is 500 usec).\n"
-"  -p, --pid=PID              The the ID of the process to which Austin should\n"
-"                             attach.\n"
-"  -s, --sleepless            suppress idle samples.\n"
-"  -?, --help                 Give this help list\n"
-"      --usage                Give a short usage message\n"
-"  -V, --version              Print program version\n"
-"\n"
-"Mandatory or optional arguments to long options are also mandatory or optional\n"
-"for any corresponding short options.\n"
-"\n"
-"Report bugs to <https://github.com/P403n1x87/austin/issues>.\n";
-
-static const char * usage_msg = \
-"Usage: austin [-aes?V] [-i n_usec] [-p PID] [--alt-format] [--exclude-empty]\n"
-"            [--interval=n_usec] [--pid=PID] [--sleepless] [--help] [--usage]\n"
-"            [--version] command [ARG...]\n";
-
-static arg_option opts[] = {
-  {
-    "interval",     'i', 1
-  },
-  {
-    "alt-format",   'a', 0
-  },
-  {
-    "exclude-empty",'e', 0
-  },
-  {
-    "sleepless",    's', 0
-  },
-  {
-    "pid",          'p', 1
-  },
-  {
-    "help",         '?', 0
-  },
-  {
-    "usage",        -1,  0
-  },
-  {
-    "version",      'V', 0
-  },
-  {0, 0, 0}
-};
-
-int cb(const char opt, const char * arg) {
-  pid_t l_pid;
-
-  switch (opt) {
-  case 'i':
-    if (strtonum((char *) arg, (long *) &t_sampling_interval) == 1 || t_sampling_interval < 0) {
-      printf(usage_msg);
-      return ARG_INVALID_VALUE;
-    }
-    break;
-
-  case 'a':
-    format = (char *) SAMPLE_FORMAT_ALTERNATIVE;
-    break;
-
-  case 'e':
-    exclude_empty = 1;
-    break;
-
-  case 's':
-    sleepless = 1;
-    break;
-
-  case 'p':
-    if (strtonum((char *) arg, (long *) &l_pid) == 1 || l_pid <= 0) {
-      printf(usage_msg);
-      return ARG_INVALID_VALUE;
-    }
-    attach_pid = (pid_t) l_pid;
-    break;
-
-  case '?':
-    printf(help_msg);
-    exit(0);
-
-  case 'V':
-    printf(PROGRAM_NAME " " VERSION);
-    exit(0);
-
-  case -1:
-    printf(usage_msg);
-    exit(0);
-
-  default:
-    printf(usage_msg);
-    return ARG_STOP_PARSING;
-  }
-
-  return ARG_CONTINUE_PARSING;
-}
-
-#endif
-
-
 // ---- MAIN ------------------------------------------------------------------
-
-// ----------------------------------------------------------------------------
-static int
-parse_args(int argc, char ** argv) {
-  #if defined(__linux__)
-  struct argp args = {options, parse_opt, "command [ARG...]", doc};
-  return argp_parse(&args, argc, argv, 0, 0, 0);
-
-  #elif defined(_WIN32) || defined(_WIN64)
-  exec_arg = arg_parse(opts, cb, argc, argv) - 1;
-  return exec_arg;
-  #endif
-}
 
 // ----------------------------------------------------------------------------
 int main(int argc, char ** argv) {
   int code = 0;
 
-  parse_args(argc, argv);
+  int exec_arg = parse_args(argc, argv);
 
   if (exec_arg == 0 && attach_pid == 0)
     return -1;
